+++ pom.xml
@@ -9,10 +9,10 @@
 		<version>2.7.0</version>
 		<relativePath /> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>digital.nogran.api</groupId>
-	<artifactId>digitalnogran</artifactId>
+	<groupId>com.digitalnogran</groupId>
+	<artifactId>email</artifactId>
 	<version>1.0.0-SNAPSHOT</version>
-	<name>digital.nogran.api</name>
+	<name>email</name>
 	<description>Digital Nogran</description>
 	<properties>
 		<java.version>17</java.version>
