+++ src/main/java/com/digital/nogran/api/DigitalnogranApplication.java
@@ -1,4 +1,4 @@
-package digital.nogran.api;
+package com.digital.nogran.api;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
