+++ src/main/java/com/digital/nogran/api/service/EmailService.java
@@ -1,0 +1,9 @@
+package com.digital.nogran.api.service;
+
+import com.digital.nogran.api.model.EmailDetails;
+
+public interface EmailService {
+    String sendSimpleMail(EmailDetails details);
+
+    String sendMailWithAttachment(EmailDetails details);
+}
