--- src/main/java/digital/nogran/api/controller/ContatoController.java
@@ -1,50 +1,0 @@
-package digital.nogran.api.controller;
-
-import java.util.List;
-
-import javax.validation.Valid;
-
-import digital.nogran.api.model.Contato;
-import digital.nogran.api.repository.ContatoRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/contato")
-@CrossOrigin(origins = "*", allowedHeaders = "*")
-public class ContatoController {
-	
-	@Autowired
-	private ContatoRepository contatoRepository;
-	
-	@GetMapping
-	public ResponseEntity<List<Contato>> getAll(){
-		return ResponseEntity.ok(contatoRepository.findAll());
-	}
-	
-	@GetMapping("/{id}")
-	public ResponseEntity<Contato> getById(@PathVariable Long id){
-		return contatoRepository.findById(id)
-				.map(resposta -> ResponseEntity.ok(resposta))
-				.orElse(ResponseEntity.notFound().build());
-	}
-	
-	@GetMapping("/nome/{nome}")
-	public ResponseEntity<List<Contato>> getByNome(@PathVariable String nome){
-		return ResponseEntity.ok(contatoRepository.findAllByNomeContainingIgnoreCase(nome));
-	}
- 	
-	@PostMapping
-	public ResponseEntity<Contato> postContato(@Valid @RequestBody Contato contato){
-		return ResponseEntity.status(HttpStatus.CREATED).body(contatoRepository.save(contato));
-	}
-
-}
