--- src/main/java/digital/nogran/api/service/EmailService.java
@@ -1,16 +1,0 @@
-package digital.nogran.api.service;
-
-//Importing required classes
-import digital.nogran.api.model.EmailDetails;
-
-//Interface
-public interface EmailService {
-
- // Method
- // To send a simple email
- String sendSimpleMail(EmailDetails details);
-
- // Method
- // To send an email with attachment
- String sendMailWithAttachment(EmailDetails details);
-}
