+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.profiles.active=prod
+spring.profiles.active=dev
 
 springdoc.swagger-ui.use-root-path=true
 springdoc.api-docs.path=/v3/api-docs
