+++ src/test/java/com/digital/nogran/api/DigitalnogranApplicationTests.java
@@ -1,4 +1,4 @@
-package digital.nogran.api;
+package com.digital.nogran.api;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
